please see the above changes.
First of all Yes, we need to update suffixLink every time whenever we encounter the condition (activeLength != 0) and the input (aaabaaaaa (credit goes to @vijju123 (codechef id) ) makes it clear! Now, it's neccessary to do this after reaching the correct activeNode(after walkdown it may be changed!) so, we have to do it after walkdown.